libhardware: Bring in GPS-related code from horseshoe
Copy gps.h and hardware.h from horseshoe to replace the existing jb_3.2
code. We're copying the files and uploading a new gerrit instead of
cherrypicking corresponding gerrits because jb_3.2 seems to have changes
that we might not need and gps code needs to be consistent with what
exists on horseshoe.

Change-Id: I3da4aba4b57c2a416f95e79b526b22ef9806b4dd